perf: refractor code using react hook custom
