+++ client/src/pages/HomePage.jsx
@@ -1,29 +1,32 @@
 import { useState, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 import axios from "axios";
+import useBlogPosts from "../utility/useBlogPosts";
 
 function HomePage() {
   const navigate = useNavigate();
 
-  const [posts, setPosts] = useState([]);
-  const [isError, setIsError] = useState(null);
-  const [isLoading, setIsLoading] = useState(null);
+  // const [posts, setPosts] = useState([]);
+  // const [isError, setIsError] = useState(null);
+  // const [isLoading, setIsLoading] = useState(null);
 
-  const getPosts = async () => {
-    try {
-      setIsError(false);
-      setIsLoading(true);
-      const results = await axios("http://localhost:4000/posts");
-      setPosts(results.data.data);
-      setIsLoading(false);
-    } catch (error) {
-      setIsError(true);
-    }
-  };
+  const { posts, isError, isLoading } = useBlogPosts();
 
-  useEffect(() => {
-    getPosts();
-  }, []);
+  // const getPosts = async () => {
+  //   try {
+  //     setIsError(false);
+  //     setIsLoading(true);
+  //     const results = await axios("http://localhost:4000/posts");
+  //     setPosts(results.data.data);
+  //     setIsLoading(false);
+  //   } catch (error) {
+  //     setIsError(true);
+  //   }
+  // };
+
+  // useEffect(() => {
+  //   getPosts();
+  // }, []);
 
   return (
     <div>
